refactor: move function out of init
